wip: restructure wizard utils module
